koord-manager: mv slo-controller metrics to util
